refactor(cli): extract improve command family into cli/commands/improve.ts
