+++ apps/axctl/src/cli/index.ts
@@ -11,24 +11,16 @@ import {
     buildSkillsByRoleNext,
     buildSkillsRolesNext,
     buildRolesNext,
-    buildImproveProposalsNext,
 } from "../nav/next-links.ts";
 import { printNextLinks } from "./next-format.ts";
 import { findCommitWindow } from "@ax/lib/git-window";
 import { AxConfig } from "@ax/lib/config";
 import { orAbsent } from "@ax/lib/shared/fs-error";
 import { ProcessService } from "@ax/lib/process";
-import { prettyPrint, surrealLiteral } from "@ax/lib/json";
+import { prettyPrint } from "@ax/lib/json";
 import { prettifyProjectSlug } from "@ax/lib/shared/project-slug";
 import { AppLayer } from "@ax/lib/layers";
-import { deriveCheckpoints } from "../ingest/derive-checkpoints.ts";
 import { loadAgentScopeMap } from "../ingest/agent-scope.ts";
-import { runAgentAccept } from "../improve/agent-accept.ts";
-import { acceptProposal, rejectProposal } from "../improve/actions.ts";
-import { lintFiles } from "../improve/lint.ts";
-import { listProposals, type ProposalRow } from "../improve/list.ts";
-import { recommend, formatRecommendations, copyToClipboard, selectByIndices, parseIndexInput } from "../improve/recommend.ts";
-import { showExperiment, formatShow } from "../improve/show.ts";
 import { cmdShare } from "./share.ts";
 import { cmdStar, maybePrintStarNudge } from "./star-nudge.ts";
 import { cmdSkillsClassify } from "./skills-classify.ts";
@@ -79,8 +71,6 @@ import {
     renderAllRolesTable,
     renderAllRolesJson,
 } from "./role-format.ts";
-import { homedir } from "node:os";
-import { surrealString } from "@ax/lib/shared/surql";
 import { ingestClaudeInsights } from "../ingest/claude-insights.ts";
 // backfillInvokedPositions - Phase B will register this as invokedPositionsStage.
 import { deriveSignals } from "../ingest/derive-signals.ts";
@@ -119,6 +109,7 @@ import { costsGroupCommand, locCommand, pricingCommand, costsRuntime } from "./c
 import { recallCommand, recallRuntime } from "./commands/recall.ts";
 import { hookCommand, hooksCommand, hooksRuntime } from "./commands/hooks.ts";
 import { retroCommand, retroRuntime } from "./commands/retro.ts";
+import { improveCommand, improveRuntime } from "./commands/improve.ts";
 import type { RuntimeManifest } from "./commands/manifest.ts";
 import { resolvePwdRepository } from "../pwd.ts";
 import { detectStaleness } from "@ax/lib/transcript-staleness";
@@ -1963,600 +1954,6 @@ const classifiersCommand = Command.make("classifiers").pipe(
     ]),
 );
 
-/**
- * axctl improve - surface the experiment-loop proposal shortlist.
- * Phase C2 ships read-only `list` + `show`. accept/reject/verdict land in
- * C3/C4/C8 with the scaffold-on-accept fix that closes the
- * manual-step-dropout problem the adversarial review flagged.
- */
-const formatProposalLine = (row: ProposalRow): string => {
-    const freq = String(row.frequency).padStart(6);
-    const conf = (row.confidence ?? "").padEnd(6);
-    const status = (row.status ?? "").padEnd(10);
-    const form = (row.form ?? "").padEnd(11);
-    const sig = (row.dedupe_sig ?? "").padEnd(24);
-    return `${freq}  ${conf}  ${status}  ${form}  ${sig}  ${row.title}`;
-};
-
-const cmdImproveList = (args: string[]) =>
-    Effect.gen(function* () {
-        const json = args.includes("--json");
-        const limit = parsePositiveIntFlag("improve list", "limit", args, 30);
-        const formFilter = flag("form", args);
-        const statusFilter = flag("status", args) ?? "open";
-        const rows = yield* listProposals({
-            status: statusFilter,
-            ...(formFilter !== undefined ? { form: formFilter } : {}),
-            limit,
-        });
-        if (json) {
-            console.log(prettyPrint(rows));
-            return;
-        }
-        const improveNext = buildImproveProposalsNext(
-            rows.map((r) => ({ sig: r.dedupe_sig, title: r.title })),
-        );
-        if (rows.length === 0) {
-            console.log("(no proposals match filter)");
-            printNextLinks(improveNext);
-            return;
-        }
-        printNextLinks(improveNext);
-        console.log(`  freq  conf    status      form         dedupe_sig                title`);
-        for (const row of rows) console.log(formatProposalLine(row));
-    });
-
-const cmdImproveShow = (args: string[]) =>
-    Effect.gen(function* () {
-        const json = args.includes("--json");
-        const positional = args.filter((a) => !a.startsWith("--"))[0];
-        if (positional === undefined) {
-            console.error("axctl improve show: missing <id> (use a dedupe_sig from `axctl improve list`)");
-            process.exit(2);
-        }
-        const result = yield* showExperiment({ sigOrId: positional });
-        if (json) {
-            console.log(prettyPrint(result));
-            return;
-        }
-        if (result === null) {
-            process.stderr.write(`no proposal matched ${positional}\n`);
-            process.exit(2);
-        }
-        console.log(formatShow(result));
-    });
-
-const cmdImproveLint = (args: string[]) =>
-    Effect.gen(function* () {
-        const json = args.includes("--json");
-        const staleDays = parsePositiveIntFlag("improve lint", "stale-days", args, 7);
-        // Collect --root values (repeatable)
-        const roots: string[] = [];
-        for (const a of args) {
-            if (a.startsWith("--root=")) roots.push(a.slice("--root=".length));
-        }
-        const report = yield* lintFiles({
-            ...(roots.length > 0 ? { roots } : {}),
-            staleDays,
-        });
-        if (json) {
-            console.log(JSON.stringify(report, null, 2));
-        } else {
-            for (const f of report.errors) {
-                console.log(`error  ${f.rule}: ${f.message} (${f.path})`);
-            }
-            for (const f of report.warnings) {
-                console.log(`warn   ${f.rule}: ${f.message} (${f.path})`);
-            }
-            for (const f of report.infos) {
-                console.log(`info   ${f.rule}: ${f.message} (${f.path})`);
-            }
-            for (const r of report.reconciled) {
-                const suffix = r.taskDeleted ? ` (removed ${r.taskDeleted})` : "";
-                console.log(`reconciled ${r.shortId}: ${r.previousStatus} -> ${r.nextStatus}${suffix}`);
-            }
-            const allEmpty =
-                report.errors.length === 0 &&
-                report.warnings.length === 0 &&
-                report.infos.length === 0 &&
-                report.reconciled.length === 0;
-            if (allEmpty) console.log("clean.");
-        }
-        if (report.errors.length > 0) {
-            process.exit(2);
-        } else if (report.warnings.length > 0) {
-            process.exitCode = 1;
-        }
-    });
-
-const cmdImproveRecommend = (args: string[]) =>
-    Effect.gen(function* () {
-        const json = args.includes("--json");
-        const noClipboard = args.includes("--no-clipboard");
-        const apply = args.includes("--apply");
-        const limit = parsePositiveIntFlag("improve recommend", "limit", args, 5);
-        const sinceDays = parseOptionalPositiveIntFlag("improve recommend", "since", args);
-        // Collect --form values (repeatable; also tolerate comma-separated)
-        const forms: string[] = [];
-        for (const a of args) {
-            if (a.startsWith("--form=")) {
-                const val = a.slice("--form=".length);
-                for (const f of val.split(",").map((s) => s.trim()).filter((s) => s.length > 0)) {
-                    forms.push(f);
-                }
-            }
-        }
-        const items = yield* recommend({
-            limit,
-            ...(forms.length > 0 ? { forms } : {}),
-            ...(sinceDays === undefined ? {} : { sinceDays }),
-        });
-        if (json) {
-            console.log(JSON.stringify(items, null, 2));
-            return;
-        }
-        printNextLinks(
-            buildImproveProposalsNext(
-                items.map((i) => ({ sig: i.shortId, title: i.title })),
-            ),
-        );
-        const formatted = formatRecommendations(items);
-        console.log(formatted);
-        if (items.length > 0 && !noClipboard) {
-            const copied = copyToClipboard(formatted);
-            if (copied) console.log("\n[copied to clipboard]");
-        }
-        if (apply && items.length > 0) {
-            // Print numbered list for reference
-            process.stdout.write("\n");
-            items.forEach((item, i) => {
-                process.stdout.write(`  ${i + 1}. ${item.shortId}  ${item.title}\n`);
-            });
-            process.stdout.write(`\nPick indices to accept (e.g. \`1 3\` or \`1-3\`): `);
-            const input = yield* Effect.promise(
-                () =>
-                    new Promise<string>((resolve) => {
-                        process.stdin.once("data", (b) => {
-                            resolve(b.toString().trim());
-                            process.stdin.pause();
-                        });
-                        process.stdin.resume();
-                    }),
-            );
-            const picked = selectByIndices(items, parseIndexInput(input, items.length));
-            for (const item of picked) {
-                const result = yield* acceptProposal({ sigOrId: item.shortId });
-                const taskSuffix = result.task_path ? ` -> ${result.task_path}` : "";
-                console.log(`${item.shortId}: ${result.status}${taskSuffix}`);
-            }
-        }
-    });
-
-const improveRecommendCommand = Command.make(
-    "recommend",
-    {
-        limit: Flag.integer("limit").pipe(Flag.withDefault(5)),
-        form: Flag.string("form").pipe(Flag.atLeast(0)),
-        since: Flag.integer("since").pipe(Flag.optional),
-        json: Flag.boolean("json").pipe(Flag.withDefault(false)),
-        noClipboard: Flag.boolean("no-clipboard").pipe(Flag.withDefault(false)),
-        apply: Flag.boolean("apply").pipe(
-            Flag.withDefault(false),
-            Flag.withDescription("Interactive: pick a proposal from the printed list and accept inline (loops until you quit). Combine with --with-agent on the accept side via the prompt."),
-        ),
-    },
-    ({ limit, form, since, json, noClipboard, apply }) =>
-        cmdImproveRecommend([
-            `--limit=${limit}`,
-            ...[...form].flatMap((f) => [`--form=${f}`]),
-            ...intArg("since", optionValue(since)),
-            ...boolArg("json", json),
-            ...boolArg("no-clipboard", noClipboard),
-            ...boolArg("apply", apply),
-        ]),
-).pipe(Command.withDescription("Rank open proposals by confidence × recency × frequency and print the top N as paste-ready blocks (with `<!--ax:id-->` provenance markers). --apply for interactive accept loop."));
-
-const improveLintCommand = Command.make(
-    "lint",
-    {
-        root: Flag.string("root").pipe(Flag.atLeast(0)),
-        json: jsonFlag,
-        staleDays: Flag.integer("stale-days").pipe(Flag.withDefault(7)),
-    },
-    ({ root, json, staleDays }) =>
-        cmdImproveLint([
-            ...[...root].map((r) => `--root=${r}`),
-            ...boolArg("json", json),
-            `--stale-days=${staleDays}`,
-        ]),
-).pipe(Command.withDescription("Scan grounded agent files (AGENTS.md / CLAUDE.md / skills) for `<!--ax:id-->` markers, reconcile against the DB, remove consumed `.ax/tasks/<id>.md` briefs, warn on orphans + tasks older than --stale-days (default 7)."));
-
-const improveListCommand = Command.make(
-    "list",
-    {
-        limit: positiveLimit(30),
-        form: Flag.string("form").pipe(Flag.optional),
-        status: Flag.string("status").pipe(Flag.optional),
-        json: jsonFlag,
-    },
-    ({ limit, form, status, json }) =>
-        cmdImproveList([
-            `--limit=${limit}`,
-            ...stringArg("form", optionValue(form)),
-            ...stringArg("status", optionValue(status)),
-            ...boolArg("json", json),
-        ]),
-).pipe(Command.withDescription("List open experiment-loop proposals (ranked by frequency)"));
-
-const improveShowCommand = Command.make(
-    "show",
-    {
-        id: Argument.string("id"),
-        json: jsonFlag,
-    },
-    ({ id, json }) => cmdImproveShow([id, ...boolArg("json", json)]),
-).pipe(Command.withDescription("Show experiment evidence + status for one proposal id"));
-
-const cmdImproveReject = (args: string[]) =>
-    Effect.gen(function* () {
-        const positional = args.filter((a) => !a.startsWith("--"))[0];
-        if (positional === undefined) {
-            console.error("axctl improve reject: missing <id>");
-            process.exit(2);
-        }
-        const reason = flag("reason", args) ?? "not_worth_packaging";
-        const result = yield* rejectProposal({ sigOrId: positional, reason });
-        if (result.status !== "ok") {
-            console.error(result.message ?? `failed to reject proposal ${positional}`);
-            process.exit(2);
-        }
-        console.log(`proposal status -> rejected (reason: ${result.reason})`);
-    });
-
-const improveAcceptCommand = Command.make(
-    "accept",
-    {
-        id: Argument.string("id"),
-        force: Flag.boolean("force").pipe(Flag.withDefault(false)),
-        withAgent: Flag.boolean("with-agent").pipe(
-            Flag.withDefault(false),
-            Flag.withDescription("After scaffold, spawn a `claude -p` subagent to read the stub + sibling skills and rewrite SKILL.md with concrete guidance. Streams to terminal. Implies --auto-scaffold."),
-        ),
-        autoScaffold: Flag.boolean("auto-scaffold").pipe(
-            Flag.withDefault(false),
-            Flag.withDescription("Skip the `.ax/tasks/<id>.md` brief and write SKILL.md directly (skill form only). Use when you want the file now, not a brief to hand to your agent."),
-        ),
-    },
-    ({ id, force, withAgent, autoScaffold }) =>
-        Effect.gen(function* () {
-            const result = yield* acceptProposal({ sigOrId: id, force, autoScaffold });
-
-            if (result.status === "not_found") {
-                console.error(result.message ?? `no proposal matched ${id}`);
-                process.exit(2);
-            }
-            if (result.status === "wrong_status") {
-                console.error(result.message ?? "proposal already processed");
-                const ex = result.existing_experiment;
-                if (ex) {
-                    console.error(`  experiment   ${ex.id}`);
-                    if (ex.artifact_path) console.error(`  scaffold     ${ex.artifact_path}`);
-                    if (ex.scaffolded_at) console.error(`  scaffolded   ${ex.scaffolded_at}`);
-                    if (ex.locked_verdict) console.error(`  verdict      ${ex.locked_verdict}`);
-                }
-                process.exit(2);
-            }
-            if (result.status === "unsupported_form") {
-                console.error(result.message ?? "unsupported form");
-                process.exit(2);
-            }
-            if (result.status === "missing_payload") {
-                console.error(result.message ?? "missing payload");
-                process.exit(2);
-            }
-            if (result.status === "scaffold_exists") {
-                console.error(result.message ?? "scaffold already exists (use --force to overwrite)");
-                process.exit(2);
-            }
-
-            // status === "ok"
-            if (result.task_path) {
-                console.log(`task emitted at ${result.task_path}`);
-                console.log(`apply with your agent: \`claude "do ${result.task_path}"\``);
-                console.log(`reconcile after edit: \`axctl improve lint\``);
-            } else if (result.artifact_path) {
-                console.log(`scaffolded ${result.artifact_path}`);
-                console.log(`experiment ${result.experiment_id ?? ""} created`);
-                console.log(`proposal status -> accepted`);
-            }
-
-            if (withAgent && result.artifact_path && result.proposal) {
-                // Spawn the claude subagent to enrich the freshly-scaffolded SKILL.md.
-                // We call runAgentAccept directly here rather than re-entering
-                // cmdImproveAccept, which would hit the `status !== 'open'` guard
-                // (the proposal was just marked accepted above).
-                let retroSummaries: readonly string[] = [];
-                const baselineRaw = result.proposal.baseline;
-                if (typeof baselineRaw === "string" && baselineRaw.length > 0) {
-                    try {
-                        const parsed = JSON.parse(baselineRaw) as {
-                            tool?: string;
-                            sessionKeys?: unknown;
-                            frequency?: number;
-                        };
-                        if (Array.isArray(parsed.sessionKeys)) {
-                            const tool = parsed.tool ?? "tool";
-                            retroSummaries = parsed.sessionKeys
-                                .filter((s): s is string => typeof s === "string")
-                                .slice(0, 5)
-                                .map((s) => `session ${s}: top tool ${tool} failed (cluster freq=${parsed.frequency ?? "?"})`);
-                        }
-                    } catch {
-                        // ignore - baseline shape may evolve
-                    }
-                }
-                console.log("");
-                console.log("spawning claude subagent to enrich the stub…");
-                const agentResult = yield* runAgentAccept({
-                    skillPath: result.artifact_path!,
-                    proposalTitle: result.proposal!.title,
-                    hypothesis: result.proposal!.hypothesis,
-                    triggerPattern: result.proposal!.triggerPattern ?? "",
-                    proposedBehavior: result.proposal!.proposedBehavior,
-                    retroSummaries,
-                    relatedSkillsDir: process.env.AX_SKILLS_SCAFFOLD_DIR ?? `${homedir()}/.claude/skills`,
-                });
-                if (agentResult.skillEnriched) {
-                    console.log(`agent enriched ${result.artifact_path}`);
-                }
-                if (agentResult.planWritten && agentResult.planPath) {
-                    console.log(`agent wrote plan ${agentResult.planPath}`);
-                }
-                if (agentResult.exitCode !== 0) {
-                    console.log(`agent exit code ${agentResult.exitCode} (stub still scaffolded; experiment row unchanged)`);
-                }
-            }
-        }),
-).pipe(Command.withDescription("Accept a proposal. Default emits a `.ax/tasks/<id>.md` brief to hand to your agent (Claude Code, Codex). --auto-scaffold writes SKILL.md directly. --with-agent dispatches a subagent to enrich the stub."));
-
-const improveRejectCommand = Command.make(
-    "reject",
-    {
-        id: Argument.string("id"),
-        reason: Flag.string("reason").pipe(Flag.optional),
-    },
-    ({ id, reason }) => cmdImproveReject([id, ...stringArg("reason", optionValue(reason))]),
-).pipe(Command.withDescription("Reject a proposal (dedupe blocks future re-proposal of same trigger)"));
-
-const ALLOWED_VERDICTS: ReadonlySet<string> = new Set([
-    "adopted", "ignored", "regressed", "partial", "no_longer_needed",
-]);
-
-/**
- * `axctl improve verdict` - surface checkpoint-derived suggested verdicts
- * for each active experiment, let the human lock the final one. Three modes:
- *
- *   axctl improve verdict
- *     -> tabular listing of every experiment with its newest suggested verdict
- *
- *   axctl improve verdict <id>
- *     -> drill-down for one experiment + each checkpoint snapshot
- *
- *   axctl improve verdict <id> --set <verdict>
- *     -> writes user_verdict on the most recent checkpoint AND
- *        locks experiment.locked_verdict so future derive runs stop scoring
- */
-const cmdImproveVerdict = (args: string[]) =>
-    Effect.gen(function* () {
-        const positional = args.filter((a) => !a.startsWith("--"))[0];
-        const setIdx = args.findIndex((a) => a === "--set" || a.startsWith("--set="));
-        let setValue: string | undefined;
-        if (setIdx >= 0) {
-            const raw = args[setIdx];
-            if (raw.startsWith("--set=")) setValue = raw.slice("--set=".length);
-            else setValue = args[setIdx + 1];
-        }
-        const json = args.includes("--json");
-        const db = yield* SurrealClient;
-        if (positional === undefined) {
-            const rows = yield* db.query<[Array<Record<string, unknown>>]>(
-                `SELECT
-                    proposal.title AS title,
-                    proposal.dedupe_sig AS dedupe_sig,
-                    artifact_path,
-                    type::string(created_at) AS created_at,
-                    type::string(scaffolded_at) AS scaffolded_at,
-                    locked_verdict,
-                    (SELECT kind, suggested, user_verdict, type::string(observed_at) AS observed_at FROM checkpoint WHERE experiment = $parent.id ORDER BY observed_at DESC LIMIT 1)[0] AS latest_checkpoint
-                FROM experiment ORDER BY created_at DESC LIMIT 30;`,
-            );
-            const list = rows?.[0] ?? [];
-            if (json) { console.log(prettyPrint(list)); return; }
-            if (list.length === 0) {
-                console.log("(no experiments yet - accept a proposal first via `axctl improve accept <sig>`)");
-                return;
-            }
-            console.log("Current experiments (newest first):");
-            for (const row of list) {
-                const cp = row.latest_checkpoint as Record<string, unknown> | null;
-                const verdict = row.locked_verdict
-                    ? `[locked: ${String(row.locked_verdict)}]`
-                    : cp
-                        ? `[${String(cp.kind ?? "?")} suggested: ${String(cp.suggested ?? "?")}]`
-                        : "[no checkpoint yet]";
-                console.log(`  ${String(row.dedupe_sig ?? "?")}  ${verdict}  ${String(row.title ?? "?")}`);
-            }
-            console.log("");
-            console.log("Run `axctl improve checkpoint` to refresh due windows.");
-            console.log("Run `axctl improve verdict <sig> --set <verdict>` to lock.");
-            return;
-        }
-        const idLiteral = surrealLiteral(positional);
-        const sel = yield* db.query<[Array<Record<string, unknown>>]>(
-            `SELECT
-                id,
-                proposal.title AS title,
-                proposal.dedupe_sig AS dedupe_sig,
-                proposal.status AS proposal_status,
-                artifact_path,
-                type::string(created_at) AS created_at,
-                type::string(scaffolded_at) AS scaffolded_at,
-                locked_verdict,
-                (SELECT id, kind, suggested, user_verdict, measured, type::string(observed_at) AS observed_at FROM checkpoint WHERE experiment = $parent.id ORDER BY observed_at DESC) AS checkpoints
-            FROM experiment
-            WHERE proposal.dedupe_sig = ${idLiteral} OR id = ${idLiteral}
-            LIMIT 1;`,
-        );
-        const row = (sel?.[0] ?? [])[0];
-        if (!row) {
-            console.error(`no experiment matched ${positional} (check \`axctl improve list --status=accepted\`)`);
-            process.exit(2);
-        }
-        const checkpoints = (row.checkpoints as Array<Record<string, unknown>> | undefined) ?? [];
-
-        if (setValue !== undefined) {
-            if (!ALLOWED_VERDICTS.has(setValue)) {
-                console.error(`--set must be one of: ${[...ALLOWED_VERDICTS].sort().join(", ")}`);
-                process.exit(2);
-            }
-            if (row.locked_verdict) {
-                console.error(`experiment already locked: ${String(row.locked_verdict)}`);
-                process.exit(2);
-            }
-            const experimentId = String(row.id ?? "");
-            const latestCp = checkpoints[0];
-            const stmts: string[] = [
-                `UPDATE ${experimentId} SET locked_verdict = ${surrealString(setValue)};`,
-            ];
-            if (latestCp?.id) {
-                stmts.push(`UPDATE ${String(latestCp.id)} SET user_verdict = ${surrealString(setValue)};`);
-            }
-            yield* db.query(stmts.join(""));
-            console.log(`verdict locked: ${setValue}`);
-            return;
-        }
-
-        if (json) { console.log(prettyPrint(row)); return; }
-        console.log(`${String(row.title ?? "?")}`);
-        console.log(`  dedupe_sig    ${String(row.dedupe_sig ?? "?")}`);
-        console.log(`  experiment    ${String(row.id ?? "?")}`);
-        console.log(`  status        ${String(row.proposal_status ?? "?")}`);
-        console.log(`  artifact      ${String(row.artifact_path ?? "(none)")}`);
-        console.log(`  scaffolded_at ${String(row.scaffolded_at ?? "(none)")}`);
-        console.log(`  verdict       ${row.locked_verdict ? String(row.locked_verdict) + " (locked)" : "pending"}`);
-        if (checkpoints.length === 0) {
-            console.log(`  checkpoints   none (run \`axctl improve checkpoint\` once due windows pass)`);
-        } else {
-            console.log(`  checkpoints:`);
-            for (const cp of checkpoints) {
-                const measured = cp.measured as Record<string, unknown> | string | undefined;
-                const opp = (typeof measured === "object" && measured) ? Number(measured.opportunities ?? 0) : 0;
-                const add = (typeof measured === "object" && measured) ? Number(measured.addressed ?? 0) : 0;
-                console.log(
-                    `    ${String(cp.kind ?? "?")}  observed=${String(cp.observed_at ?? "?")}  ` +
-                    `opportunities=${opp} addressed=${add}  suggested=${String(cp.suggested ?? "?")}  ` +
-                    `user_verdict=${cp.user_verdict ? String(cp.user_verdict) : "(none)"}`,
-                );
-            }
-        }
-        console.log("");
-        console.log(`Lock the verdict: \`axctl improve verdict ${String(row.dedupe_sig ?? "<sig>")} --set <verdict>\``);
-        console.log(`  verdicts: ${[...ALLOWED_VERDICTS].sort().join(", ")}`);
-    });
-
-const improveVerdictCommand = Command.make(
-    "verdict",
-    {
-        id: Argument.string("id").pipe(Argument.optional),
-        set: Flag.string("set").pipe(Flag.optional),
-        json: jsonFlag,
-    },
-    ({ id, set, json }) => {
-        const idValue = optionValue(id);
-        const setValue = optionValue(set);
-        return cmdImproveVerdict([
-            ...(idValue === undefined ? [] : [idValue]),
-            ...stringArg("set", setValue),
-            ...boolArg("json", json),
-        ]);
-    },
-).pipe(Command.withDescription("Show experiment verdict state; --set adopted|ignored|regressed|partial|no_longer_needed locks it"));
-
-/**
- * `axctl improve reset --yes` - drop every experiment-loop row in DB.
- *
- * Destructive. Used by UAT to start from a clean slate before re-running
- * the full propose -> accept -> verdict flow. Wipes the 9 experiment-loop
- * tables in dependency order; underlying evidence (friction_event,
- * skill_candidate, etc) is left alone so re-derivation can rebuild
- * proposals against the same signal.
- */
-const cmdImproveReset = (args: string[]) =>
-    Effect.gen(function* () {
-        if (!args.includes("--yes")) {
-            console.error("axctl improve reset: refusing to wipe without --yes");
-            console.error("  drops: checkpoint, opportunity, experiment, cites_evidence,");
-            console.error("         skill_proposal, subagent_proposal, hook_proposal,");
-            console.error("         guidance_proposal, automation_proposal, proposal");
-            process.exit(2);
-        }
-        const db = yield* SurrealClient;
-        // Dependency order: checkpoint -> opportunity -> experiment ->
-        // cites_evidence -> per-form payloads -> proposal. Relations cascade
-        // via REFERENCE ON DELETE CASCADE on the schema, but we delete
-        // bottom-up to keep this explicit + auditable.
-        yield* db.query(`
-            DELETE checkpoint;
-            DELETE opportunity;
-            DELETE experiment;
-            DELETE cites_evidence;
-            DELETE skill_proposal;
-            DELETE subagent_proposal;
-            DELETE hook_proposal;
-            DELETE guidance_proposal;
-            DELETE automation_proposal;
-            DELETE proposal;
-        `);
-        console.log("experiment-loop state cleared. Run \`ax ingest --stages=proposals,opportunities\` to rebuild.");
-    });
-
-const improveResetCommand = Command.make(
-    "reset",
-    {
-        yes: Flag.boolean("yes").pipe(Flag.withDefault(false)),
-    },
-    ({ yes }) => cmdImproveReset([...boolArg("yes", yes)]),
-).pipe(Command.withDescription("Wipe all experiment-loop state (proposals/experiments/checkpoints). Requires --yes."));
-
-const cmdImproveCheckpoint = (args: string[]) =>
-    Effect.gen(function* () {
-        const force = args.includes("--force");
-        const json = args.includes("--json");
-        const stats = yield* deriveCheckpoints({ force });
-        if (json) {
-            console.log(prettyPrint(stats));
-            return;
-        }
-        console.log(`checkpoints scanned: ${stats.experimentsScanned} experiments`);
-        console.log(`checkpoints inserted: ${stats.checkpointsInserted}`);
-        console.log(`checkpoints skipped: ${stats.checkpointsSkipped}`);
-        if (stats.checkpointsInserted === 0) {
-            console.log("");
-            console.log("No new windows due. Re-run with --force to refresh existing checkpoints");
-            console.log("(use `axctl improve verdict <id>` to see suggested verdicts).");
-        }
-    });
-
-const improveCheckpointCommand = Command.make(
-    "checkpoint",
-    {
-        force: Flag.boolean("force").pipe(Flag.withDefault(false)),
-        json: jsonFlag,
-    },
-    ({ force, json }) =>
-        cmdImproveCheckpoint([...boolArg("force", force), ...boolArg("json", json)]),
-).pipe(Command.withDescription("Compute checkpoint snapshots at +3/+10/+30 sessions for active experiments (session-count windows, not calendar days - see issue #83)"));
-
 // ---------------------------------------------------------------------------
 // ax sessions - windowed session queries (F2, F3)
 // ---------------------------------------------------------------------------
@@ -3260,21 +2657,6 @@ const sessionsCommand = Command.make("sessions").pipe(
     ]),
 );
 
-const improveCommand = Command.make("improve").pipe(
-    Command.withDescription("Experiment loop: rank proposals (recommend), accept (emit task brief or scaffold + dispatch subagent), lint grounded agent files, track verdicts at +3/+10/+30 sessions after accept."),
-    Command.withSubcommands([
-        improveRecommendCommand,
-        improveLintCommand,
-        improveListCommand,
-        improveShowCommand,
-        improveAcceptCommand,
-        improveRejectCommand,
-        improveCheckpointCommand,
-        improveVerdictCommand,
-        improveResetCommand,
-    ]),
-);
-
 const searchCommand = Command.make(
     "search",
     {
@@ -3786,7 +3168,6 @@ const LEGACY_RUNTIME: RuntimeManifest = {
     "derive-intents": "db",
     classifiers: "db",
     sessions: "db",
-    improve: "db",
     skills: "db",
     roles: "db",
     agents: "db",
@@ -3806,6 +3187,7 @@ export const RUNTIME_BY_COMMAND: RuntimeManifest = {
     ...recallRuntime,
     ...hooksRuntime,
     ...retroRuntime,
+    ...improveRuntime,
 };
 
 // Commands whose handlers reach into SurrealClient via AppLayer (or the
